server browsing: play folder contents as a playlist instead of individual items
Subtitles support missing so far

+++ Sources/VLCPlaybackController.h
@@ -96,7 +96,7 @@ currentMediaHasTrackToChooseFrom:(BOOL)currentMediaHasTrackToChooseFrom
 - (void)setNeedsMetadataUpdate;
 - (void)scheduleSleepTimerWithInterval:(NSTimeInterval)timeInterval;
 
-- (void)playMediaList:(VLCMediaList *)mediaList firstIndex:(int)index;
+- (void)playMediaList:(VLCMediaList *)mediaList firstIndex:(NSInteger)index;
 - (void)playURL:(NSURL *)url successCallback:(NSURL*)successCallback errorCallback:(NSURL *)errorCallback;
 - (void)playURL:(NSURL *)url subtitlesFilePath:(NSString *)subsFilePath;
 - (void)remoteControlReceivedWithEvent:(UIEvent *)event;
